requirements.txt
Flask
spotipy
pandas
gunicorn
opencv-python
numpy
Pillow
tensorflow
matplotlib
